refactor: move rate models to a separate directory
