+++ contracts/InterestRateModels/TwoKinksInterestRateModel.sol
@@ -2,8 +2,8 @@
 pragma solidity 0.8.25;
 
 import { TimeManagerV8 } from "@venusprotocol/solidity-utilities/contracts/TimeManagerV8.sol";
+import { EXP_SCALE, MANTISSA_ONE } from "../lib/constants.sol";
 import { InterestRateModel } from "./InterestRateModel.sol";
-import { EXP_SCALE, MANTISSA_ONE } from "./lib/constants.sol";
 
 /**
  * @title TwoKinksInterestRateModel
