+++ tests/hardhat/InterestRateModels/JumpRateModelV2.ts
@@ -4,10 +4,10 @@ import BigNumber from "bignumber.js";
 import chai from "chai";
 import { ethers } from "hardhat";
 
-import { BSC_BLOCKS_PER_YEAR } from "../../helpers/deploymentConfig";
-import { convertToUnit } from "../../helpers/utils";
-import { AccessControlManager, JumpRateModelV2 } from "../../typechain";
-import { getDescription } from "./util/descriptionHelpers";
+import { BSC_BLOCKS_PER_YEAR } from "../../../helpers/deploymentConfig";
+import { convertToUnit } from "../../../helpers/utils";
+import { AccessControlManager, JumpRateModelV2 } from "../../../typechain";
+import { getDescription } from "../util/descriptionHelpers";
 
 const { expect } = chai;
 chai.use(smock.matchers);
